Update FrameEntryType to account for when players are defeated
