+++ src/stats/conversions.ts
@@ -1,6 +1,6 @@
 import _ from "lodash";
-import { PostFrameUpdateType } from "../types";
-import { FrameEntryType, FramesType, MoveLandedType, ConversionType, PlayerIndexedType } from "./common";
+import { FrameEntryType, FramesType, PostFrameUpdateType } from "../types";
+import { MoveLandedType, ConversionType, PlayerIndexedType } from "./common";
 import { isDamaged, isGrabbed, calcDamageTaken, isInControl, didLoseStock, Timers } from "./common";
 import { StatComputer } from "./stats";
